requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.66
langchain-text-splitters==0.3.8
langsmith==0.4.4
openai==1.93.0
orjson==3.10.18
packaging==24.2
praw==7.8.1
prawcore==2.4.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
schedule==1.2.2
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
update-checker==0.18.0
urllib3==2.5.0
websocket-client==1.8.0
zstandard==0.23.0
